refactor: extract page components from fat route files to features
all routes/*.tsx are now uniform 5-line stubs — createFileRoute + one
import. moved AIGeneratePage, JobsPage, MonitoringPage, and SettingsPage
into their respective features/ directories to match the pattern already
established by ai.tsx, analyze.tsx, and the other thin stubs.
